// generated/prisma/models/Booking.ts
export type BookingSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  stripeChargeId?: boolean
  barbershopId?: boolean
  serviceId?: boolean
  userId?: boolean
  date?: boolean
  cancelledAt?: boolean
  createdAt?: boolean
  updatedAt?: boolean
  barbershop?: boolean | Prisma.BarbershopDefaultArgs<ExtArgs>
  service?: boolean | Prisma.BarbershopServiceDefaultArgs<ExtArgs>
  user?: boolean | Prisma.UserDefaultArgs<ExtArgs>
}, ExtArgs["result"]["booking"]>
// ...
export type BookingOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<"id" | "stripeChargeId" | "barbershopId" | "serviceId" | "userId" | "date" | "cancelledAt" | "createdAt" | "updatedAt", ExtArgs["result"]["booking"]>
export type BookingInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  barbershop?: boolean | Prisma.BarbershopDefaultArgs<ExtArgs>
  service?: boolean | Prisma.BarbershopServiceDefaultArgs<ExtArgs>
  user?: boolean | Prisma.UserDefaultArgs<ExtArgs>
}
// ...
export type $BookingPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Booking"
  objects: {
    barbershop: Prisma.$BarbershopPayload<ExtArgs>
    service: Prisma.$BarbershopServicePayload<ExtArgs>
    user: Prisma.$UserPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    stripeChargeId: string | null
    barbershopId: string
    serviceId: string
    userId: string
    date: Date
    cancelledAt: Date | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["booking"]>
  composites: {}
}

export type BookingGetPayload<S extends boolean | null | undefined | BookingDefaultArgs> = runtime.Types.Result.GetResult<Prisma.$BookingPayload, S>
// ...
export type BookingDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the Booking
   */
  select?: Prisma.BookingSelect<ExtArgs> | null
  /**
   * Omit specific fields from the Booking
   */
  omit?: Prisma.BookingOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.BookingInclude<ExtArgs> | null
}